Basic Integration
Setup basic integration, about to ty some major changes by moving overlayRef into the dragging item
