+++ documents/diary/gestures.md
@@ -1,5 +1,8 @@
 # 7-30-24
 Previously, I had a setup where the equipment items would automatically begin dragging everytime a gesture was put over it. However, this would lead to problems with swiping across the scrollview. Especially as I want to add tap functionality and make the equipment items function similar to android or ios apps, I tried to use longPresses and a modified panResponder to get the job done. Unfortunately, panResponder isn't very flexible with layered responders, and therefore I am going to try react-native-gesture-handler.
 
-## React-native-gesture-handler
+## React-native-gesture-handler minimal setup
 I first started using a basic setup no a temporary screen called TestSwap. For this screen, I simply implemented a scrollview with items that took a long press to start dragging, and could respond to tap events. React-native-gesture-handler actually made the functionality much easier than expected, and it didn't take too long to figure it out. The biggest timesaver was the states emitted by the handlers, especially 'active'. By using the active state from long press, I could ensure the user kept their finger on the item for the specified time before starting the panresponder (no more need to use a timeout).
+
+## Integrating RNGH
+Integrating wasn't difficult, though I learned that I needed to specify "worklet" for passed functions called by Gesture. Basically, if I run Gestures on the UI thread, the functions called by the gesture also needs to run on the UI thread.
